Update paasta status to use http api when needed
This makes it so paasta status will force use of the HTTP api for k8s
instances whilst still defaulting to the SSH api for other instance
types. As we add features to the HTTP api e.g. for marathon app status
we can migrate it over to using the HTTP by adding to
HTTP_ONLY_INSTANCE_TYPES.